fix: Fix enabled status of up/down/top/bottom buttons on move item up/down (#165)
I don't see an open issue for this, but it seems like the logic to
{en,dis}able the move buttons is not correct when items are moved
up/down (it is correct on move to top/bottom as this causes a page
refresh). This PR fixes the button status on both newly moved items and
simplifies the overall logic a bit.

Notably - I'm not really great with web languages in general, but my
best guess for why this was broken is that the `querySelector` for
`tbody` was first matching the table for user-provided attributes (ex.
shoe size, shirt size, etc.). So my changes for the enable/disable logic
itself may not be fully necessary.

Closes #168 and probably closes #152

---------

Co-authored-by: Wingy <git@wingysam.xyz>