feat: Make price column on Wishlist table sortable (#140)
I agree with @peskyAdmin that a sortable table feature would be really
helpful (#137). This is simple implementation but seems to work pretty
nicely:


![image](https://github.com/user-attachments/assets/6c490f77-c189-461f-802b-577a0e9cf3e9)

As an aside - I've no idea what the README change is about. It seems to
be caused by the contributors workflow.

Probably closes #137 if accepted. And also probably closes #53

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Wingy <git@wingysam.xyz>