Allow annotations on forks and only auto-fix issues for pushes to master (#200)
An attempt to fixup the pipeline to allow merges.

Not sure of the target of the lint action as I've never seen it run, but
I might also suggest to set `on.[action].paths` to only run the workflow
when necessary. I'll check it out if I can get it to run.