feat: Set (shorten) animation duration for wishlist move actions (#166)
Quick PoC for shortening the animation duration of moving wishlist
items.

Frankly, I think the default duration for `zoomIn` and `zoomOut` are
downright silly. I found this 0.45s value to be reasonable through
testing - I personally prefer more like `0.33s` but I figured that may
be a little fast for some.

Would be ideal if this can be tuned in config or through environment
(pending config page impl) - any ideas?